Refactored Generator and Validator to be thread safe
